requirements.txt
python-dotenv==0.13.0
flask==1.1.2
flask-cors==3.0.8
pycaret==1.0.0
numpy==1.18.4
gunicorn==20.0.4
Jinja2==2.11.2
certifi==2020.4.5.1
itsdangerous==1.1.0